@remotion/canvas-capture: Add package
