+++ packages/canvas-capture/README.md
@@ -1,0 +1,7 @@
+# @remotion/canvas-capture
+
+Capture HTML-in-canvas content as a video
+
+## Usage
+
+This is an internal package and has no documentation.
