+++ packages/canvas-capture/bundle.ts
@@ -1,0 +1,35 @@
+import path from 'path';
+import {build} from 'bun';
+
+if (process.env.NODE_ENV !== 'production') {
+	throw new Error('This script must be run using NODE_ENV=production');
+}
+
+console.time('Generated.');
+const output = await build({
+	entrypoints: ['src/index.tsx'],
+	naming: '[name].mjs',
+	external: [
+		'mediabunny',
+		'remotion',
+		'remotion/no-react',
+		'react',
+		'react/jsx-runtime',
+		'react/jsx-dev-runtime',
+		'react-dom',
+	],
+});
+
+if (!output.success) {
+	console.log(output.logs.join('\n'));
+	process.exit(1);
+}
+
+for (const file of output.outputs) {
+	const str = await file.text();
+	const out = path.join('dist', 'esm', file.path);
+
+	await Bun.write(out, str);
+}
+
+console.timeEnd('Generated.');
