+++ packages/canvas-capture/tsconfig.json
@@ -1,0 +1,16 @@
+{
+	"extends": "../tsconfig.settings.json",
+	"compilerOptions": {
+		"rootDir": "src",
+		"outDir": "dist",
+		"jsx": "react-jsx",
+		"skipLibCheck": true,
+		"target": "ES2022",
+		"module": "es2020",
+		"moduleResolution": "bundler",
+		"declaration": true,
+		"emitDeclarationOnly": true
+	},
+	"include": ["src"],
+	"references": [{"path": "../core"}]
+}
