+++ packages/create-video/src/list-of-remotion-packages.ts
@@ -8,6 +8,7 @@ export const listOfRemotionPackages = [
 	'@remotion/bugs',
 	'@remotion/brand',
 	'@remotion/bundler',
+	'@remotion/canvas-capture',
 	'@remotion/cli',
 	'@remotion/cloudrun',
 	'@remotion/codex-plugin',
