+++ packages/example/package.json
@@ -40,6 +40,7 @@
 		"@remotion/animation-utils": "workspace:*",
 		"@remotion/babel-loader": "workspace:*",
 		"@remotion/bundler": "workspace:*",
+		"@remotion/canvas-capture": "workspace:*",
 		"@remotion/effects": "workspace:*",
 		"@remotion/captions": "workspace:*",
 		"@remotion/cli": "workspace:*",
