+++ packages/studio-shared/src/package-info.ts
@@ -7,6 +7,7 @@ export const packages = [
 	'bugs',
 	'brand',
 	'bundler',
+	'canvas-capture',
 	'cli',
 	'cloudrun',
 	'codex-plugin',
@@ -136,6 +137,7 @@ export const descriptions: {[key in Pkgs]: string | null} = {
 	core: 'Make videos programmatically',
 	lambda: 'Render Remotion videos on AWS Lambda',
 	bundler: 'Bundle Remotion compositions using Webpack',
+	'canvas-capture': 'Capture HTML-in-canvas content as a video',
 	'studio-server': 'Run a Remotion Studio with a server backend',
 	'install-whisper-cpp': 'Helpers for installing and using Whisper.cpp',
 	'whisper-web': 'Helpers for using Whisper.cpp in browser using WASM',
@@ -238,6 +240,7 @@ export const installableMap: {[key in Pkgs]: boolean} = {
 	bugs: false,
 	brand: false,
 	bundler: false,
+	'canvas-capture': false,
 	cli: false,
 	cloudrun: true,
 	'codex-plugin': false,
@@ -336,6 +339,7 @@ export const apiDocs: {[key in Pkgs]: string | null} = {
 	core: 'https://www.remotion.dev/docs/remotion',
 	lambda: 'https://www.remotion.dev/docs/lambda',
 	bundler: 'https://www.remotion.dev/docs/bundler',
+	'canvas-capture': null,
 	'lambda-client': null,
 	'serverless-client': null,
 	'studio-server': null,
