+++ packages/studio/bundle.ts
@@ -5,6 +5,7 @@ const external = [
 	'remotion',
 	'react-dom',
 	'react',
+	'@remotion/canvas-capture',
 	'@remotion/media-utils',
 	'@remotion/studio-shared',
 	'@remotion/timeline-utils',
