+++ packages/studio/package.json
@@ -26,6 +26,7 @@
 	"dependencies": {
 		"semver": "7.5.3",
 		"remotion": "workspace:*",
+		"@remotion/canvas-capture": "workspace:*",
 		"@remotion/player": "workspace:*",
 		"@remotion/media-utils": "workspace:*",
 		"@remotion/renderer": "workspace:*",
